Module cmn::constants

The constants module contains the Constants structure, which provides a collection of constant values that are used throughout the library.

Structs

Constant

Contains several commonly used mathematical and cryptographic constants.

Constants

The Constants structure holds mathematical and hash constants.

Enums

ConstantValue

Enum to represent the different constant values.

Constants

AVOGADRO

Avogadro’s constant Approximately 6.02214076 x 10^23

BOLTZMANN

Boltzmann’s constant Approximately 1.380648 x 10^-23

EULER

The base of the natural logarithm, Euler’s number (e). e ≈ 2.7182818284590452353602874713527

GAMMA

The mathematical constant γ or the Euler–Mascheroni constant. It is the limit of the difference between the harmonic series and the natural logarithm of the natural numbers.

HASH_ALGORITHM

The hash algorithm used. The default is Blake3.

HASH_COST

The cost of the hash algorithm. The default is 8.

HASH_LENGTH

The hash length is the length of the hash in bytes.

PHI

The mathematical constant φ or the golden ratio. It is the limit of the ratio of consecutive Fibonacci numbers. Φ = (1+√5)/2 = 2.cos(π/5). Diagonal of a unit-side pentagon.

PI

The mathematical constant π or the ratio of a circle’s circumference to its diameter.

PLANCK

The Planck constant, h.

SILVER_RATIO

The mathematical constant `δs’ or the silver ratio (or silver mean). δs = 1+√2. One of the silver means (n+sqrt(n2+1))/2 for n>0.

SPECIAL_CHARS

A set of special characters.

SQRT2

The mathematical constant √2 or the Pythagora’s constant or the square root of 2. It is the diagonal of a square with unit side length. 2 = 1 + √2

SQRT3

The mathematical constant √3 or the principal square root of 3. It is the length of the side of an equilateral triangle with unit side length. 3 = 1 + √3

SQRT5

The mathematical constant √5 or the principal square root of 5. It is the length of the diagonal of a regular pentagon with unit side length. 5 = 2 + 2√5

TAU

The mathematical constant τ or the ratio of a circle’s circumference to its radius.