cmn
Module constants

The constants module contains the Constants structure, which provides a collection of constant values that are used throughout the library.

Structs

Constant

Contains several commonly used mathematical and cryptographic constants.

Constants

The Constants structure holds mathematical and hash constants.

Enums

ConstantValue

Enum to represent the different constant values.

Constants

APERY

Apéry’s constant, which is the sum of the reciprocals of the positive cubes. ζ(3) ≈ 1.2020569032

AVOGADRO

Avogadro’s constant, which is the number of constituent particles contained in one mole of a substance. N_A ≈ 6.02214076 x 10^23 mol^-1

BOLTZMANN

Boltzmann’s constant, which relates the average kinetic energy of particles in a gas with the temperature of the gas. k_B ≈ 1.380648 x 10^-23 J K^-1

CATALAN

Catalan’s constant, which is the sum of the alternating harmonic series. C ≈ 0.915965594177219

COULOMB

Coulomb’s constant, which is the proportionality constant in Coulomb’s law. k_e ≈ 8.9875517923 x 10^9 N m^2 C^-2

EULER

The base of the natural logarithm, Euler’s number. e ≈ 2.7182818284590452353602874713527

FARADAY

Faraday constant, which represents the amount of electric charge carried by one mole of electrons. F ≈ 96485.33212 C mol^-1

GAMMA

The Euler-Mascheroni constant, which is the limiting difference between the harmonic series and the natural logarithm. γ ≈ 0.5772156649015329

GAS_CONSTANT

The gas constant, which relates the energy scale to the temperature scale in the ideal gas law. R ≈ 8.314462618 J mol^-1 K^-1

GLAISHER_KINKELIN

Glaisher-Kinkelin constant, which arises in the asymptotic expansion of the Barnes G-function. A ≈ 1.2824271291

GRAVITATIONAL_CONSTANT

The gravitational constant, which is the proportionality constant in Newton’s law of universal gravitation. G ≈ 6.67430 x 10^-11 m^3 kg^-1 s^-2

HASH_ALGORITHM

The hash algorithm used. The default is Blake3.

HASH_COST

The cost of the hash algorithm. The default is 8.

HASH_LENGTH

The hash length is the length of the hash in bytes.

KHINCHIN

Khinchin’s constant, which appears in the theory of continued fractions. K ≈ 2.6854520010

PHI

The golden ratio, which is the limit of the ratio of consecutive Fibonacci numbers. φ = (1 + √5) / 2 ≈ 1.6180339887498948482045868343656

PI

The ratio of a circle’s circumference to its diameter. π ≈ 3.14159265358979323846264338327950288

PLANCK

Planck’s constant, which relates the energy of a photon to its frequency. h ≈ 6.62607015 x 10^-34 J s

PLANCK_REDUCED

Planck’s reduced constant, which is Planck’s constant divided by 2π. ħ = h / (2π) ≈ 1.054571817 x 10^-34 J s

SILVER_RATIO

The silver ratio, which is one of the silver means. δ_s = 1 + √2 ≈ 2.4142135623730950488016887242097

SPECIAL_CHARS

A set of special characters.

SPEED_OF_LIGHT

The speed of light in vacuum. c ≈ 299792458 m s^-1

SQRT2

The square root of 2. √2 ≈ 1.4142135623730950488016887242097

SQRT3

The square root of 3. √3 ≈ 1.7320508075688772935274463415059

SQRT5

The square root of 5. √5 ≈ 2.23606797749979

TAU

The circle constant, which is the ratio of a circle’s circumference to its radius. τ = 2π ≈ 6.28318530717958647692528676655900577

VACUUM_PERMEABILITY

The vacuum permeability, which relates magnetic induction to magnetic field strength. μ_0 ≈ 1.25663706212 x 10^-6 N A^-2

VACUUM_PERMITTIVITY

The vacuum permittivity, which relates electric displacement to electric field strength. ε_0 ≈ 8.8541878128 x 10^-12 F m^-1